axion_lab/pyproject.toml
[project]
name = "axion-lab"
version = "0.0.1"
description = "Artifact-first Experiment Evaluation System"
readme = "README.md"
authors = [{ name = "elda27", email = "kaz.birdstick@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.35",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "psycopg2-binary>=2.9.0",
    "boto3>=1.35.0",
    "aioboto3>=13.0.0",
    "google-cloud-storage>=2.14.0",
    "httpx>=0.28.0",
    "python-ulid>=3.0.0",
    "python-dotenv>=1.0.0",
    "pytest-asyncio>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-boto3>=1.35.0",
]

[project.scripts]
axion-lab = "axion_lab.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/axion_lab", "src/axion_lab_server"]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["./src/axion_lab_alembic/**"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "DTZ", "T10", "ISC"]
ignore = ["ISC001"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = ["pytest>=9.0.2", "ruff>=0.15.0", "ty>=0.0.16"]
google = ["google-cloud-storage>=3.9.0"]
